+++ src/lib/ui/Forum.svelte
@@ -6,7 +6,7 @@
 	import ForumItems from '$lib/ui/ForumItems.svelte';
 	import {getApp} from '$lib/ui/app';
 
-	const {api, ui, data, socket} = getApp();
+	const {api, ui, socket} = getApp();
 
 	export let community: Community;
 	export let space: Space;
@@ -35,7 +35,7 @@
 		}
 	};
 
-	$: files = $data.filesBySpace[space.space_id] || [];
+	$: files = $ui.filesBySpace[space.space_id] || [];
 </script>
 
 <div class="forum">
