+++ src/lib/ui/app.ts
@@ -1,7 +1,6 @@
 import {setContext, getContext} from 'svelte';
 import type {Writable} from 'svelte/store';
 
-import type {DataStore} from '$lib/ui/data';
 import type {UiStore} from '$lib/ui/ui';
 import type {Api} from '$lib/ui/api';
 import type {SocketStore} from '$lib/ui/socket';
@@ -10,7 +9,6 @@ import type {SocketStore} from '$lib/ui/socket';
 
 export interface AppStores {
 	api: Api;
-	data: DataStore;
 	ui: UiStore;
 	socket: SocketStore;
 	devmode: Writable<boolean>;
